Added `isOpen` state
isOpen state is a state which is true if the state of the dropdown is active, else it is false. This will be used to handle callback functions more effectively incase user wants to select multiple inputs, as the normal callback of `value` will be triggered on each select/deselect, `isOpen` can be used to minimize excess computation.